Purify msh monad stack
Decouples the core from the universe in two major steps.

First, StateT is replaced with ReaderT. Updating the shell Context is no
longer possible from within the MshAction monad, which will ensure that
such stateful operations are kept to a minimum.

Second, the transformer stack MshAction is now generic in the inner
monad type. IO no longer being implicit makes it possible to write code
that is guaranteed to be pure.