Define basic commands in interactive prompt
The only command introduced is `echo`, with a dummy parser that is not
designed to scale as the project continues. However, it provides a first
proof of concept, including a foundation for reporting errors to the
user.

The changes to the main shell loop, however, should handle future
development very well and only require further refinement as part of
module refactoring.