Split shell Context into immutable and mutable parts
The Context type now only contains the data referring to immutable shell
properties set during initialization (such as the location of the
profile script). Data which is mutable over the shell's lifetime (such
as the prompt string) is now held in the Settings type.

This also adds a StateT transformer to the monad stack to manage
stateful effects modifying these Settings.